Create login module and move jwt auth stuff there
