+++ packages/server/src/login/login.module.ts
@@ -1,7 +1,19 @@
 import { Module } from '@nestjs/common';
 import { LoginController } from './login.controller';
+import { JwtStrategy } from '../login/jwt.strategy';
+import { JwtModule } from '@nestjs/jwt';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+
 
 @Module({
-  controllers: [LoginController]
+  imports: [JwtModule.registerAsync({
+    imports: [ConfigModule],
+    inject: [ConfigService],
+    useFactory: async (configService: ConfigService) => ({
+      secret: configService.get('JWT_SECRET'),
+    }),
+  })],
+  controllers: [LoginController],
+  providers: [JwtStrategy],
 })
 export class LoginModule {}
