+++ packages/server/src/profile/profile.controller.ts
@@ -1,33 +1,26 @@
 import {
   Controller,
   Get,
-  Res,
-  Query,
   UseGuards,
   Patch,
   Body,
 } from '@nestjs/common';
-import { JwtService } from '@nestjs/jwt';
-import { Response } from 'express';
 import { Connection } from 'typeorm';
 import { UserModel } from './user.entity';
 import { pick } from 'lodash';
 import { GetProfileResponse, UpdateProfileParams } from '@template/common';
-import { JwtAuthGuard } from './jwt-auth.guard';
+import { JwtAuthGuard } from '../login/jwt-auth.guard';
 import { User } from './user.decorator';
 import { NotificationService } from '../notification/notification.service';
-import { ConfigService } from '@nestjs/config';
 
 @Controller('profile')
+@UseGuards(JwtAuthGuard)
 export class ProfileController {
   constructor(
     private connection: Connection,
-    private jwtService: JwtService,
     private notifService: NotificationService,
-    private configService: ConfigService,
   ) {}
 
-  @UseGuards(JwtAuthGuard)
   @Get()
   async get(
     @User(['courses', 'courses.course', 'phoneNotif']) user: UserModel,
@@ -57,7 +50,6 @@ export class ProfileController {
     };
   }
 
-  @UseGuards(JwtAuthGuard)
   @Patch()
   async patch(
     @Body() userPatch: UpdateProfileParams,
@@ -92,15 +84,5 @@ export class ProfileController {
       courses,
       phoneNumber: user.phoneNotif?.phoneNumber,
     };
-  }
-
-  // TODO handle the khoury flow for real.
-  @Get('/entry')
-  enterFromKhoury(@Res() res: Response, @Query('userId') userId: number): void {
-    const token = this.jwtService.sign({ userId });
-    const isSecure = this.configService.get('NODE_ENV') === 'production';
-    res
-      .cookie('auth_token', token, { httpOnly: true, secure: isSecure })
-      .redirect(302, '/');
   }
 }
