+++ packages/server/src/profile/profile.module.ts
@@ -1,22 +1,11 @@
 import { Module } from '@nestjs/common';
 import { ProfileController } from './profile.controller';
-import { JwtModule } from '@nestjs/jwt';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { JwtStrategy } from './jwt.strategy';
 import { NotificationModule } from '../notification/notification.module';
 
 @Module({
   imports: [
-    JwtModule.registerAsync({
-      imports: [ConfigModule],
-      inject: [ConfigService],
-      useFactory: async (configService: ConfigService) => ({
-        secret: configService.get('JWT_SECRET'),
-      }),
-    }),
     NotificationModule,
   ],
   controllers: [ProfileController],
-  providers: [JwtStrategy],
 })
 export class ProfileModule {}
